Update patch set 1
Patch Set 1:

I'm probably willing to let feature this slide into rc1, but once we cut rc1 in a week or so it'll be too late.

If we want it in 2.15 then it should probably be developed in stable-2.15 instead of master, though.

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>