Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> https://imgur.com/a/FnUPD Updated

"Diff against" implies that the left value is the actual patchNum, whereas the right value is the base (at least to me). Is this the case? If not, I think we may want to reconsider the language.

Patch-set: 5
Reviewer: Gerrit User 1029297 <1029297@173816e5-2b9a-37c3-8a2e-48639d4f1153>