Update patch set 1
Patch Set 1:

> I'm not sure why this is necessary on stable-2.15? Was this a bug?

Yep, because of the recent work done to the router (gr-router) if the url is not specified in there, then it results in a 404. So all links with external bypasses that and allows the url to work.

Patch-set: 1
Reviewer: Gerrit User 1029953 <1029953@173816e5-2b9a-37c3-8a2e-48639d4f1153>