Update patch set 3
Patch Set 3:

> I think this should be backported to 2.14 as it includes fixes for edge and chrome.

I don't think this should go to 2.14. Although PG was introduced in 2.14 it was described as "still in development" in the release notes.

PG is described in the 2.15 release notes as "mature enough for most uses", so I think if we backport this anywhere it should only be to stable-2.15.

Patch-set: 3
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>