Update patch set 9
Patch Set 9:

When I ran the tests locally, CreateProjectIT#createSameProjectFromTwoConcurrentRequests failed for me. It reported a NullPointerException at com.google.gerrit.server.index.project.ProjectIndexerImpl.index(ProjectIndexerImpl.java:68). That resulted in the failed test assertion: "Not true that <[500, 409]> contains all of <[201, 409]>. It is missing <[201]>".

When I re-ran the test, it succeeded.

Patch-set: 9
CC: Gerrit User 1030912 <1030912@173816e5-2b9a-37c3-8a2e-48639d4f1153>