Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > Does this work when NoteDb is write-only? I forget exactly, but I thought that ChangeUpdate expects the ChangeNotes you pass to it to have an accurate meta ref ID.
> > 
> > Will check this ASAP.
> 
> Do you expect to fix this soon?  I'm just wondering if I should wait for this before releasing 2.14.5.1

We have deployed 2.14 and we're having some problems so I cannot prioritize this in the next two days. I'll work on it though on the weekend, so for Monday I'll have an answer to Dave's concerns.

Patch-set: 1
Reviewer: Gerrit User 1019328 <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>