Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> I wonder should we backport this to 2.15?

There are a few changes (this one, and one button organization one @wyatt worked on) that are not straightforward cherrypicks. I think they should get in, but I'm not sure the best way to do that. If we cherrypick/resolve conflicts there may be new conflicts introduced when merging 2.15 back into master, and I want to minimize that.

Patch-set: 8