Update patch set 2
Patch Set 2: Code-Review-1

Would it be worthwhile to leave this restriction in? It would make it less likely for users to get in the situation where they have permissions on these refs that are not respected.

Patch-set: 2
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1