Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Does this still require accounts to be a reviewer (as opposed to CC)?

Yes, unregistered entities can't be reviewers as they could never vote.

When building this feature we thought about adding a conversion if a user that was added as unregistered CC would later register (just convert them) but we dropped that to keep the feature simple and also because it wasn't part of the feature request.

Patch-set: 2
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>