Update patch set 3
Patch Set 3: Code-Review-1

I think strictly speaking this is a backwards incompatible change to the REST API.

The documentation of DETAILED_ACCOUNTS says that the email and username fields are included in resulting AccountInfos, in other words, users can assume those fields being non-null if they pass this option. After this change, that assumption no longer holds.

I am undecided whether it's ok to break the API in this way, but I'm leaning towards no. If we agree it is ok, then at a minimum we need to change the documentation.

If it's not ok, then we could still keep this change, we would just have to manually populate those fields with some placeholders, rather than relying on the values returned from an empty AccountState. But honestly, at that point I'm not sure it's worth it.

Patch-set: 3
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1