Update patch set 3
Patch Set 3: Code-Review+2

This is essentially the same as the previous one that already had +2.  This is merged from an earlier revision on stable-2.13 that omits the version upgrade, and hence the only difference in this merge is the conflict resolution was not needed.

Patch-set: 3
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+2