Update patch set 2
Patch Set 2:

(1 comment)

> Would it make sense to apply this whole series to 2.15 and merge it
 > back into master?

False negatives in diff calculation is one of the most annoying issues. And rebase operation is something done very often. That why I'm in favor of applying this series to stable-2.15 branch.

Sure, doing that would invalidate the whole IntralineDiff cache, but that is something that needs to happen anyway.

Patch-set: 2
CC: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>