Update patch set 1
Patch Set 1:

Some findings from our offline discussion for other interested readers:

We'll make sure the scenario of having a hidden project show up in ListProjects (given the options) and setting it back to ACTIVE or READ_ONLY over the API works.

All other usages of ACCESS in this series will not receive the isOwner() check in addition so that we don't add a lot of boiler-plate to the code base for a feature that isn't used very often.

Patch-set: 1
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>