Update patch set 2
Patch Set 2:

> 
> Does it really need to be done on stable-2.14? I understand, when we fix critical bugs on very stable branches < 2.15 or even introduce new features that cannot wait until 2.16/3.0, but doing such invasive refactoring only for adding a new test (or even extend existing tests to new sub-system) seems questionable to me.

We want to get Elasticsearch working on the stable-2.14 branch so that we can introduce it for 2.14.8.

As you know, most of the ES implementation is already there, so we're not really adding a new feature.

Some issues were found, and the fixes we've done for those have been mostly isolated in the ES implementation so should not have any impact on the rest of the system.

The remaining work has been to extend the base query tests to add missing coverage for search predicates.  This improves coverage for both Lucene and ES, and as a result we have also uncovered a couple of issues that affect both systems.

Test coverage for reindex with ES is now the only missing part before we can say that ES has the same test coverage as Lucene.

So from that perspective I think it's acceptable to do it on the stable branch.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>