Update patch set 1
Patch Set 1:

This is something of an RFC at the moment.

I wanted to be able to add more fields that only apply to persistent caches[1], without cluttering CacheProvider up with precondition checks that only apply to persistent caches. In addition, in the past, both cache(...) and persist(...) returned an object of type CacheBinding, and the more persistent-only methods we have in CacheBinding, the more likely someone would call one of them by accident on a non-persistent cache.

OTOH, we now have a rather complex inheritance structure for CacheProvider, which I'm not very happy with.

[1] Specifically: version, key converter, value converter

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Work-in-progress: false