Update patch set 3
Patch Set 3:

I am on the fence if this is a good idea or not. It seems that previously we wanted to hide the put methods from the extension API by using this PrivateInternals_ subclass construct.

Did you run into any issues with the other implementations you are working on? Or is this change a general clean up?

Patch-set: 3
Reviewer: Gerrit User 1026112 <1026112@173816e5-2b9a-37c3-8a2e-48639d4f1153>