Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Good point. Although I would say "logically the same" means "are the same account" and we only need to compare the Id to prove that.  Would it be sufficient to document this caveat in the equals method's javadoc?
> 
> Yes, that would be a solution. However, the Java library methods that are calling equals() with the canonical meaning would still fail :-( 

Right, but that's the point of this change (or rather the following one that throws UnsupportedOperationException when equals is called on something other than IdentifiedUser.

> Shall we just call it "equalsIds()" ?

That's one way, but it's basically the same as what you did in the other change on stable-2.15 that is since abandoned.  It doesn't solve the problem of callers using the == and != operators, and it's less intuitive than a method named equals.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>