Update patch set 1
Patch Set 1: Code-Review-1

I'm not really ok with implementing equals and hashCode for only a subset of the CurrentUser hierarchy. If we do it for one subclass, IMO we need to do it properly for the whole hierarchy. It's not an easy problem[1], but a partial implementation can be worse than no implementation.

Also, even if we were to just implement this for IdentifiedUser, I'm not convinced that comparing only account ID is always correct. Two issues off the top of my head (there may be more):

* If two IdentifiedUsers have different remote peers, should they be considered equal?
* If two IdentifiedUsers have different AccountStates because the account was updated and AccountCache entry got evicted between the times when they were created, should they be considered equal?

If the answer is that the use cases we care about are only interested in comparing account IDs, then perhaps we could get away with adding a single boolean hasSameAccountId(CurrentUser other) method. That name makes it clear that we are not interested in a holistic definition of equality, only in this one bit.

[1] http://www.angelikalanger.com/Articles/JavaSolutions/SecretsOfEquals/Equals-2.html

Patch-set: 1
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1