Update patch set 5
Patch Set 5:

> I think you can do
> 
>  git reset --soft 668944bc88
>  git commit -c bf04ff88c8
>  git push origin HEAD:refs/for/master

Oof, looks like I picked the wrong base. Sorry for the mess.

I think you'd want

  git reset -- soft 49b63e03
  git commit -c bf04ff88c8
  git push origin HEAD:refs/for/master

Do you run "git pull" as part of your workflow? That runs a fetch followed by a merge and may be the source of some of your problems. "git pull --rebase" may or may not be easier to use (personally, I tend to not pull at all, which is even simpler).

Patch-set: 5