Update patch set 1
Patch Set 1: Code-Review-1

Please update the commit message to mention the new features in 5.0, and more importantly, that it introduces a lot of deprecations.

See [1] which I did to upgrade to an earlier RC, and the rest of the series [2] which works on removing the deprecations.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/182310
[2] https://gerrit-review.googlesource.com/q/topic:%22jgit-5.0.0%22+status:open

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1