Update patch set 9
Patch Set 9:

Stepping in for Wyatt here -- I like this change quite a bit, but I have one main issue. We need the gr-change-list element (and all of its respective children) to be completely API-agnostic, and so the addition of gr-rest-api-interface to the element is a no-go for me. 

I'd recommend following the approach in I3b2ecfe26. All that has to change here is the addition of an event that's caught by the parent components and makes the correct API call, as opposed to the change-list making the API call directly.

Patch-set: 9
Reviewer: Gerrit User 1029297 <1029297@173816e5-2b9a-37c3-8a2e-48639d4f1153>