Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> + UX designers: Do you have an opinion on whether WIP changes should show up on non-author's dashboards if they are assigned or CC'ed to them? My intuition is that WIP should indicate that the CL is not ready for a full review/merge, and thus should not be picked up by contributors just browsing their dashboard. However, if the author is seeking explicit early advice on a WIP change from a specific person and indicates so by assigning them, or if someone wants to follow a WIP change somebody has pointed them to even while it is in WIP mode, that should be possible. 
> Réda seems to prefer to never show WIP changes on non-authors dashboard. 
> I can live with either decision, but this seems like a UX question (and depending on the decision we may have to adjust help tooltips etc.).

I can live with either decision too.
But if WIP changes were to be kept visible in the assignee dashboard, it is very important to us that they do not appear as if an action from the assignee was required (no bold for WIP in the assignee dashboard).

Patch-set: 4
Reviewer: Gerrit User 1057319 <1057319@173816e5-2b9a-37c3-8a2e-48639d4f1153>