Update patch set 4
Patch Set 4:

I think a change can be marked WIP even after starting the review ie., if the author gets a bunch of feedback on patchset one and wants to change the code based on that, they can mark the change as WIP. If we stop showing a change when it has been marked WIP, esp if a non-author has already taken a look at the change once, it will appear to the non-author as if the change has randomly disappeared from the dashboard. We should avoid doing that.

Interaction Designer

arnab

banerjee

Patch-set: 4
Tag: mailMessageId=<CAA_3c+6zTEuYpg0RDCUwi7ReDiWvj0oDo39baqfa==0p3M8sXQ@mail.gmail.com>