Update patch set 1
Abandoned

While this appears to work in live, a lot of testcases are reliant on the current functionality. I had looked at a few testcases that this touched and they had TODO's marked on "not being sent from Owner", but it is clear that I need to go more in depth and spend more time to solve this and there's no point in this just sitting as a private change until then.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon