LabelType: make its construtor private
"LabelType" is a value class which would be better
if it's an @AutoValue class.

Since "LabelType" is a relatively big class, this
adaption will be done step by step to make sure
it's a no-op for all callers.

This commit moves the first step toward direction
by initializing all "LabelType" through static
factory methods, e.g. #create. As a result we could
mark its constructor private.

Change-Id: Ib880da23de381525517e90a5596bdc73e228ac84