LabelType: rename "getRefPatterns" to "refPatterns"
This commit is part of the commits adapting "LabelType"
to be an @AutoValue class.

In the end, when "LabelType" becomes an @AutoValue,
"refPatterns" will be accessed through "#refPatterns"
provided by the @AutoValue class. Renaming it in this
commit to make review easier.

This commit also moves several methods to group their
"getter" and "setter" together.

Change-Id: I9b256c9d978e446d990595456348235bba0c6dcf