Update patch set 6
Patch Set 6:

Not yet. However, when we would to it, we would not be able to use bazel action cache. Can you test a gerrit build with `bazelisk build :release` on CI, how long would that take? We could set remote bazel cache, that we started to work on during the hackathon, but this is also not ready yet.

Patch-set: 6
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>