Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Interesting... see the last message sent by this merge...
> The mail subject is different than the change subject... is this a bug? I expect the mail subject to follow the change subject, in our case it is the subject of the first change in the review.
> 
> Subject: Change in gerrit/gerrit[master]: Upgrade gitiles to 0.3-1
> <snip>
> --0000000000006be336058c51fa1e
> Content-Type: text/plain; charset="UTF-8"; format=flowed; delsp=yes
> 
> Marco Miller has submitted this change and it was merged. (  
> https://gerrit-review.googlesource.com/c/gerrit/+/227208 )
> 
> Change subject: Upgrade gitiles to 0.3-2

This is intentional. The template is configured to always show the original change subject in the email subject.  See [1].  This of course breaks when the change destination is moved, because then the subject changes if the template includes the branch name in the subject.

When [1] was implemented it was not possible to move the change, so that wasn't a consideration.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/62737

Patch-set: 6