Update patch set 3
Patch Set 3:

(1 comment)

What happens, if the client side change is released before the server side change? If that is not a problem, then document that in the commit message. It it is a problem, then separate the changes into client and server and make sure that the client change is only submitted after the server change is released.

Patch-set: 3
Reviewer: Gerrit User 1013302 <1013302@173816e5-2b9a-37c3-8a2e-48639d4f1153>