export rules private in progresss
Each file exported also has the type of file, while the only
supported types are "SUBMODULE" and "REGULAR" while magic files
are omitted.

Also, each file has the modification type (add, delete, etc).

Add a test that ensures that the commit that adds a generic file
(in this case a.txt) either with "addition" or "deletion".

Add another test that has one submodule file, and ensure that it
is exported with "SUBMODULE" as part of the tuple, and another test
that ensures it's possible to make a rule that is true when the it
contains a submodule.

Change-Id: I960402e8508834d86db0c5691e59c5174fb0e00a