Evict group caches after periodic group reindexing
When group attributes are changed or group membership is changed, the
group caches evicted correctly. When however on replica instance
periodic group reindexing takes place the groups and the groups_byname
were missed to be evicted.  That leads to stale caches on replicas.

This could be workaround by setting maxAge attribute of these caches to
5 minutes, to match the default time that periodic group indexer is
running on the replica instance. But in that case an unnecessary
eviction of all entries would happen, also for the groups that weren't
changed.

Given that the periodic group indexer is already optimized and only
reindex a group if it's stale (by memoizing the ref states), we move
the code that evict the caches in the index code path. That way we
avoid code duplication in 3 different places:

o Create group for primary instance
o Update group for primary instance
o Periodic group indexer for replica instance

Bug: Issue 40014962
Release-Notes: Evict group caches after periodic group reindexing in
replicas
Change-Id: I7cefc9deef2bfc12ae85d7aaa1b0b8cf02256f60