Add a hidden property to account.config
The new property controls how the account in returned from the REST API
endpoints.

If hidden, account won't be returned on account queries, and all REST
API requests will only return the Account.Id (no personal information).

The hidden users will only be visible to themselves and to users with
VIEW_ALL_ACCOUNTS and MODIFY_ACCOUNT capabilities.

The hidden users will be represented as Anonymous users on UI.

This change only adds support to store this property in the
account.config The handlers to update the property and the permission
restrictions will be added in the follow up changes.

No compatibility problems are expected, since the new property is not
required in account.config or Account cache.

An auth.defaultNewAccountHidden option in gerrit.config defines the
default for the new accounts. If not set, the property is not written on
account creation. This allows to backfill the property for accounts,
once the feature is rolled out.

The property defaults to false, if not found in account.config.

The detailed documentation will be added in the follow up changes.

Change-Id: I8b2d0bec4d4841239ba325bc7b4c935bf1567ffc
Release-Notes: skip