Add a metric to measure the success of rebasing on behalf of the uploader
If the 2 person principle is enforced by using a submit requirement
expression with 'label:Code-Review=MAX,user=non_uploader' there is the
problem that changes are not submittable with an approval of a reviewer
if the reviewer rebases the change (because by rebasing the change the
reviewer becomes the uploader and then their approvals are ignored).

To solve this problem we support rebasing on behalf of the uploader now.
Rebasing a change on behalf of the uploader means that the uploader
stays intact when the reviewer rebases the change. This means an
approval of the reviewer on the rebased change is sufficient to submit
it.

To measure the success of this feature we want to know how many changes
get submitted with approvals of the rebaser, that would not have been
submittable if a normal rebase had been done.

To find out about this when submitting changes we check for changes
whichs latest patch set was created by rebase on behalf of the uploader:
1. whether any submit requirement expressions ignores approvals of the
   uploader (i.e. uses 'user=non_uploader')
2. whether any of the submit requirement expressions that ignore
   approvals of the uploader would be non-passing if approvals of the
   rebaser aka the real uploader were ignored instead of ignoring the
   approvals of the uploader

Admittedly the computation of the metric is a bit hacky and we likely
want to remove it after the support for rebasing on behalf of the
uploader has been rolled out and its success has been measured. Hence we
are not officially documenting this metric.

Release-Notes: skip
Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia5b3a3c2c74891c9e467e7aad3b9dcad174f3646