[RFC] Implement Workspaces
A workspace is a lightweight fork of a repository. The idea behind it is that
whatever happens locally in your Git repo, so actions like:
  - Creation of commits
  - Rebasing
  - Prepating changes to be uploaded

happen on the server. This has the following advantages:
  - Allow for integrations of other web tools into your immediate
    workflow (editor, linters, CI, etc).
  - Automatic syncing between different computers

A lightweight fork is a separate repository on the server that
uses the mainlines object database as 'alternate'. This means it
always writes objects to it's own ObjDb, but when performing
lookups will consider objects in mainline as well. Since this
is a Git-native concept, admins can also use Git CLI to
look into these repos on the server (e.g. for maintainance).

The ACL story is simple (though, not implemented in this change):
Workspaces are prefixed with Account.Id and only accessible by
their owner.

We'll add APIs so that users can 'upstream' their commits from
a workspace to the mainline. This will be a mere in-mem copy of
objects.

Git clients can simply be set up to auto-sync commits using a
post-commit hook.

Change-Id: Ieb2b8b59082d091510288366835f92df50ca06a4