Disable support for named queries
With BlockNamedDestinationsAndQueries we have a validator that rejects
adding named queries for accounts and groups. With this change we
disable loading existing named queries (from logging we concluded
that there are none, but it's not 100% sure) and return a proper error
message when named queries are used in change queries.

Bug: Google b/325022991
Change-Id: Ia931433aa15f991e2e5d6f71c89792a6e62e88cb
Signed-off-by: Edwin Kempin <ekempin@google.com>