Fix starred changes clash after repo import from another site
Introduce `virtualId` to change data and initiate it with the
`legacy_id_str` from the index so that changes list screen (which is
populated from the index only) could be properly updated with star bit.

Ensure that `changeServerId` is loaded from notes when notes are
available in the `ChangeData`.

Finally use the `virtualId` in all calls to get/set starred changes so
that the refs clash is avoided. As a result the starred changes ref
gets modified to the following form:

  refs/starred-changes/NN/<virtual-id>/<account-id>

where `virtual-id` is `server-id` encoded `change-id` for imported
changes and `change-id` for existing/newly created ones.

Issue (331016670) to be addressed as the follow-up: toggle star from
changes list screen doesn't work for imported changes - the problem is
that UI uses change.id to obtain the project first which clashes with
existing changes. It works fine on the single change screen.

Bug: Issue 325309573
Release-Notes: Starred changes clash when importing changes from another instance
Change-Id: I85c946efcab5d20dd25682cb758bb9601c93b1d2