Ensure plugin modules are bound in the baseInjector
It's required to do an explicit bind when using child injectors in order
to prevent just-in-time bindings from being promoted to an ancestor
injector. This is a well documented issue in Guice [1] that is only
mentioned a couple places in the javadocs, most prominently at [2].

In Gerrit this became a problem with the introduction of the plugin
ApiModule. That module is expected to be the base for other plugins, but
shouldn't get any bindings promoted from the per-plugin
ServerPluginInfoModule. Adding explicit bind() calls in
ServerPluginInfoModule for each of the plugin modules prevents that from
happening, as seen in the new PluginBindingsIT suite.

[1] https://github.com/google/guice/issues/973
[2] https://google.github.io/guice/api-docs/latest/javadoc/com/google/inject/Injector.html#createChildInjector(java.lang.Iterable)

Bug: Issue 332512505
Change-Id: I8633083d97d28ea9d1e61a90bb86efdb399c12d0
Release-Notes: Fix installing plugins injecting the injector when an ApiModule plugin is also installed