Add REST API to clean old changes
Change cleanup could be configured in the gerrit.config. Then it
would run on a configured schedule. This might not always be wanted
or could even lead to issues, if the repositories are shared by
multiple Gerrit instances, which try to run the cleanup concurrently.

This change adds a REST API endpoint to run the change cleanup
manually.

Release-Notes: Add REST API to clean old changes
Change-Id: Id04b1061561b52037b1805b05db66b37b6e1298d