Fix comment widgets losing focus by stable sorting
We have been hunting this bug for 6+ months: Mysteriously the comment
textarea would sometimes lose focus. We had lots of users complaining,
but the Gerrit team rarely ran into the problem and we did not have any
way of reliably reproducing the issue.

Today we did find a reproduction and could thus debug the issue:
<gr-diff-host> uses the `repeat` directive to render comment thread
elements. If the underlying array changes how the threads are sorted,
then the directive has to remove a comment thread element and insert
it somewhere else. This is not really user visible, but of course,
removing the element for a split second does remove the focus.

Thus the root cause is that the sorting of comment threads is not
stable. We have to make sure that *for drafts* the `updated` timestamp
is not considered for sorting. Otherwise every (auto-)save operation
would potentially change the order of the comment thread arrays. For
published comments it is fine to use the `updated` timestamp, because
it won't change anymore.

The sorting order actually does not matter at all (as far as we know).
It must just be *stable*!

Google-Bug-Id: b/299507066
Bug: Issue 295374158
Release-Notes: Fix comment widgets losing focus
Change-Id: Ia7c0b11c83f2954f1e248da894158b60e0e310d6