Record duration of performance events in nanos instead of in millis
If the duration is recorded in millis all operations that take less than
1ms are recorded with 0ms. This means they will have the exact same
start and end times because the start time is computed by (end time
minus duration), although start and end time have nanosecond precision.
This is bad when visualising events in a gantt chart as events that have
the exact same start and end time (and hence a 0 duration) cannot be
shown.  This is confusing when analysing performance logs because
expected events do not show up (it's unclear if they do not show up
because they were not executed or because they were faster than 1ms).

Notes, TimerContext already records the elapsed time in nanos and
provides them with TimeUnit.NANOSECONDS to the timers.

While we are here, make the formatting of how the elapsed time is shown
in debug logs everywhere consistent (with a space between the number and
the "ms").

Release-Notes: skip
Change-Id: I8ad36b03b4dd79a6fe975920ad7b33700bf0e87f
Signed-off-by: Edwin Kempin <ekempin@google.com>