Extract plugin load order comparator into separate class
To ensure proper plugin load order Gerrit used a private
implementation of `java.util.Comparator` interface. The responsibility
of that implementation is to ensure that:

 * plugins with `Gerrit-ApiModule` entry in `Manifest.mf` are first,
 * all other plugins are sorted in the natural order.

As this is a crucial part of plugin loading and may be extended later
(eg. by including load priority in `Manifest.mf`) it will be good to
have at least basic testing for the current logic.

This change is extracting this `Comparator` code into a standalone class
and adding test coverage for it.

Release-Notes: skip
Change-Id: I5587d40bc53ffd47049dbccda550dfa9748677ef