Allow influence plugin load order from gerrit.config
In some complex deployments of Gerrit, the order of plugins' load during
the startup time matters.

Currently, the only way to influence it is to rename plugin JAR file in
the `plugins/` directory. Which is a bit inconvenient and possibly
error-prone.

With this change, it's now possible to use `plugins.loadPriority`
option of `gerrit.config` to influence the order of loading plugins
during the startup.

The `plugins.loadPriority` is a list configuration option, that accepts
plugin names (as discovered by Gerrit from the `MANIFEST.MF` value of
`Gerrit-PluginName` or plugin JAR file in `plugins/` directory). The
earlier the plugin is defined in `gerrit.config` the sooner (or higher
priority) it will have.

Bug: Issue 335744331
Release-Notes: Change plugin load order using `plugins.loadPriority`
Change-Id: Ie2e646ff8cdea8529f79084055ae2401d22e111c