Introduce ConvertibleToProto interface
This empty interface is intended to raise the visibility on the
expectation that the implementing java entity has an equivalent proto
which should be kept in sync.

This interface is located in the `common` directory, because some of the
entities that will use it are under the `extensions.client` package
(e.g. user preferences).

Ideally, this would've been an annotation, however, it's intended to be
used as a generic parameter for a safe extension to the `ProtoConverter`
interface. (See the next change).

Release-Notes: skip
Google-Bug-Id: b/335372403
Change-Id: Ib19f91e06b0ebb2423af15f6ddd2134e3663ec30