Add a `draft-actions-end` endpoint slot to <gr-comment>
Allows adding extra content at the end of the row of draft action
buttons.

Example usage: https://imgur.com/a/UwJtcWG

Also, we are not rendering the link icon anymore when editing. This
is rather cluttering the UI and competes with other content that
might be rendered in the slot.

Release-Notes: skip
Google-Bug-Id: b/336237249
Change-Id: I0b7dbae319116a085d3a8ba5af1830d6f61b91e6