Add release war target without webfonts
This adds a new release war target that builds the Documentation without
webfonts. The ability to build the documentation without webfonts is
relatively new and I couldn't find a straightforward way to build a
release war without webfonts so I added a new target for this.

Release-Notes: Add release build target that omits webfonts
Change-Id: I247190aa677efdeb6bb1daaa4a78945f3bef88a6