Add a cache for autocompleting comments
We already had existing logic in gr-comment that would keep a hint
around, if the user typed conforming to the hint. But it is nicer
to have this in a separate class (and have tests for it).

Also, we want to support that the user may be backspacing, for
example if a suggestion popped up for a split second, but they had
already typed another character accidentally. Then backspace should
bring back the former hint immediately.

Release-Notes: skip
Google-Bug-Id: b/339170117
Change-Id: Id751dccc1747bf4ccf76836654265ed7d6cc5d6a