Fix issues with account details filling
Users have reported an issue with bogus account labels when the
account label was trying to fill in the details. One issue was that
the account (even the detailed one) does not have an `email` property
set. So the underlying `isDetailedAccount()` utility was producing
false return values and initiated the account to be looked up again
and again.

Another issue was that `Object.assign()` would actually modify the
cached object in place and thus potentially mess up the account object
for all subsequent callers.

Release-Notes: Fix bogus account labels.
Bug: Issue 340606666
Change-Id: Ib91fe57ad8026ef2a1daf61ddc2af5954ffd39ab