Add config for protecting project config from unreviewed updates.
A project config can be updated directly (without creating a change)
using gerrit rest api. In some cases we would like to ensure that all
changes to project config are properly reviewed.

This change adds a configuration for gerrit which requires creating a
change for each project config update. This allows to configure gerrit
in a way that each project's config change is reviewed (by setting a
correct permissions/submit requirements).

The implementation is based on permissions, because internally (in
google), this feature will work more like permissions (i.e. it will
depend on a user and on a project protection settings, which can be
dynamically changed).

This change doesn't update gerrit UI, it will be done in the follow up
changes.

Release-Notes: A new config for protecting project config from unreviewed updates.
Google-Bug-Id: b/330832714
Change-Id: I68190060a0146c59447577d520861baa151691df