Fix translation of SubmitRecord status to SubmitRequirement
The refactoring of the adaptation of legacy SubmitRecord(s) to
SubmitRequirement(s) made with Iccc3f3c8585 overlooked the status
and made a Prolog rule that generate a submit record with status OK
to be considered a non verified requirement.

Example, the following SubmitRecord:

{
  "status_code": 200,
  "value": {
    "status": "OK",
    "ok": {
      "Overriden": {
        "_account_id": 100000
      }
    },
    "need": {
      "Still-Needed_label": {}
    },
    "may": {
      "Owners-Override": {}
    }
  },
  "headers": {},
  "caching": {
    "type": "NONE",
    "age": 0,
    "must_revalidate": false
  }
}

Was translated to a SubmitRequiremnt of
Still-Needed_label instead of resulting into a submittable
change.

The problem was not present in the initial version of the
translation in Ie0fbf0e70a where the mapping was done properly
by the mapStatus() function:

  private static Status mapStatus(SubmitRecord submitRecord) {
    switch (submitRecord.status) {
      case OK:
      case CLOSED:
      case FORCED:
        return Status.PASS;
      case NOT_READY:
        return Status.FAIL;
      case RULE_ERROR:
      default:
        return Status.ERROR;
    }
  }

However, the refactoring overlooked the status and the current
version considers only the FORCED condition, ignoring the OK and
CLOSED.

Expose the SubmitRecor.Status.allowsSubmission() that makes the
correct check, taking into considration all the correct status
values.

Change-Id: Iea8beaf7b92b0846fe6f2de1614153e175fc2a43